requirements.txt
keras
pillow
numpy
flask